Revert "[flang][NFC] Move OpenMP related passes into a separate directory (#1…"
This reverts commit 87eeed1f0ebe57abffde560c25dd9829dc6038f3.
