+++ flang/include/flang/Optimizer/CMakeLists.txt
@@ -2,4 +2,3 @@ add_subdirectory(CodeGen)
 add_subdirectory(Dialect)
 add_subdirectory(HLFIR)
 add_subdirectory(Transforms)
-add_subdirectory(OpenMP)
