+++ flang/lib/Optimizer/CMakeLists.txt
@@ -5,4 +5,3 @@ add_subdirectory(HLFIR)
 add_subdirectory(Support)
 add_subdirectory(Transforms)
 add_subdirectory(Analysis)
-add_subdirectory(OpenMP)
