--- flang/lib/Optimizer/OpenMP/CMakeLists.txt
@@ -1,25 +1,0 @@
-get_property(dialect_libs GLOBAL PROPERTY MLIR_DIALECT_LIBS)
-
-add_flang_library(FlangOpenMPTransforms
-  FunctionFiltering.cpp
-  MapInfoFinalization.cpp
-  MarkDeclareTarget.cpp
-
-  DEPENDS
-  FIRDialect
-  HLFIROpsIncGen
-  FlangOpenMPPassesIncGen
-
-  LINK_LIBS
-  FIRAnalysis
-  FIRBuilder
-  FIRCodeGen
-  FIRDialect
-  FIRDialectSupport
-  FIRSupport
-  FortranCommon
-  MLIRFuncDialect
-  MLIROpenMPDialect
-  HLFIRDialect
-  MLIRIR
-)
