+++ flang/lib/Optimizer/Transforms/OMPFunctionFiltering.cpp
@@ -1,4 +1,4 @@
-//===- FunctionFiltering.cpp -------------------------------------------===//
+//===- OMPFunctionFiltering.cpp -------------------------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -13,7 +13,7 @@
 
 #include "flang/Optimizer/Dialect/FIRDialect.h"
 #include "flang/Optimizer/Dialect/FIROpsSupport.h"
-#include "flang/Optimizer/OpenMP/Passes.h"
+#include "flang/Optimizer/Transforms/Passes.h"
 
 #include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/Dialect/OpenMP/OpenMPDialect.h"
@@ -21,18 +21,18 @@
 #include "mlir/IR/BuiltinOps.h"
 #include "llvm/ADT/SmallVector.h"
 
-namespace flangomp {
-#define GEN_PASS_DEF_FUNCTIONFILTERING
-#include "flang/Optimizer/OpenMP/Passes.h.inc"
-} // namespace flangomp
+namespace fir {
+#define GEN_PASS_DEF_OMPFUNCTIONFILTERING
+#include "flang/Optimizer/Transforms/Passes.h.inc"
+} // namespace fir
 
 using namespace mlir;
 
 namespace {
-class FunctionFilteringPass
-    : public flangomp::impl::FunctionFilteringBase<FunctionFilteringPass> {
+class OMPFunctionFilteringPass
+    : public fir::impl::OMPFunctionFilteringBase<OMPFunctionFilteringPass> {
 public:
-  FunctionFilteringPass() = default;
+  OMPFunctionFilteringPass() = default;
 
   void runOnOperation() override {
     MLIRContext *context = &getContext();
