+++ flang/tools/bbc/CMakeLists.txt
@@ -25,7 +25,6 @@ FIRTransforms
 FIRBuilder
 HLFIRDialect
 HLFIRTransforms
-FlangOpenMPTransforms
 ${dialect_libs}
 ${extension_libs}
 MLIRAffineToStandard
