+++ flang/tools/fir-opt/CMakeLists.txt
@@ -19,7 +19,6 @@ target_link_libraries(fir-opt PRIVATE
   FIRCodeGen
   HLFIRDialect
   HLFIRTransforms
-  FlangOpenMPTransforms
   FIRAnalysis
   ${test_libs}
   ${dialect_libs}
