+++ flang/tools/tco/CMakeLists.txt
@@ -17,7 +17,6 @@ target_link_libraries(tco PRIVATE
   FIRBuilder
   HLFIRDialect
   HLFIRTransforms
-  FlangOpenMPTransforms
   ${dialect_libs}
   ${extension_libs}
   MLIRIR
